Fixes to catalog file generation
The next/previous links in the index nodes were incorrect -- they need
to point to the nodes in each level of the tree, not to ones in the
global order.

Also fix directory thread records for the root directory (the key
needs to be 7 bytes long instead of 6) and skip ones for files (it
results in incorrect alias resolving -- they don't appear to be
load bearing).